# Reduce allocations in `san.jl` by using `@views`